chore: restructure old usage services
